requirements.txt
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==5.3.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
fastjsonschema==2.19.1
filelock==3.12.2
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.6
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.6
notebook==6.5.6
notebook-shim==0.2.4
numpy==1.21.6
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.43
prov==2.0.0
psutil==5.9.8
pycparser==2.21
pydot==2.0.0
pygments==2.17.2
PyMuPDF==1.22.5
pyparsing==3.1.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
pywin32==306
pywinpty==2.0.10
pyxnat==1.6.2
pyzmq==24.0.1
rdflib==6.3.2
requests==2.31.0
scipy==1.7.3
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
traits==6.3.2
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
zipp==3.15.0
